RDKB-62322 : rdkconfig_get call is failing to return password
Reason for change: Mesh initialization begins only after all certificates
are procured and ready for use. During startup, Mesh calls rdkconfig_get
to retrieve the password needed to decrypt the private key. However, on
some field devices this call fails, causing OpenSync to repeatedly
attempt—and fail—to connect to the controller.

This fix introduces a mechanism to log the failure to a file. The SelfHeal
script will monitor the presence of this file and, if detected, will
trigger a restart of OpenSync to recover from the failure.

OVS does not dynamically reload certificates or keys or wont try to get
password while running. This is a design behavior.

Test Procedure: Log in self heal
Risks: Medium
Priority: P1

Signed-off-by: Mohammed Arif Mannoppilly Hassan <arifcep@gmail.com>